move `slots` module
