+++ rasa/shared/core/slots.py
