+++ tests/shared/core/test_slots.py
@@ -1,7 +1,7 @@
 import pytest
 
 from rasa.core.constants import DEFAULT_CATEGORICAL_SLOT_VALUE
-from rasa.core.slots import (
+from rasa.shared.core.slots import (
     Slot,
     TextSlot,
     BooleanSlot,
